% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_string}
\alias{clean_string}
\title{Clean and pad an encoded shortened h3 integer.}
\usage{
clean_string(encoded_short_int)
}
\description{
This creates a placekey without formatting.
}
\keyword{internal}
