% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_placekey.R
\name{construct_sg_query}
\alias{construct_sg_query}
\title{Construct GraphQL query from vector of placekeys}
\usage{
construct_sg_query(placekeys)
}
\description{
Construct GraphQL query from vector of placekeys
}
\keyword{internal}
