% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coord_to_h3_int}
\alias{coord_to_h3_int}
\title{Return the raw 64 bit inter H3 index from a latitude, longitude and resolution}
\usage{
coord_to_h3_int(lat, lon, res)
}
\arguments{
\item{lat}{latitude in degrees}

\item{lon}{longitude in degrees}

\item{resolution}{resolution, higher is more granular}
}
\description{
Return the raw 64 bit inter H3 index from a latitude, longitude and resolution
}
