% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_to_placekey.R
\name{coord_to_placekey}
\alias{coord_to_placekey}
\title{Generate Placekey from a coordinate}
\usage{
coord_to_placekey(lat, long)
}
\arguments{
\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}
}
\description{
Generate Placekey from a coordinate
}
\examples{
coord_to_placekey(29.76328, -95.36327)
#> "@675-9z6-5nw-m2"
}
