% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode_short_int}
\alias{encode_short_int}
\title{Encode a shortened h3 integer into the placekey alphabet}
\usage{
encode_short_int(short_h3_int)
}
\description{
Encode a shortened h3 integer into the placekey alphabet
}
\keyword{internal}
