% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h3_int_to_string}
\alias{h3_int_to_string}
\title{h3_int_to_string takes an h3 integer representation and returns the string format.}
\usage{
h3_int_to_string(index)
}
\arguments{
\item{index}{as an integer representation.}
}
\description{
h3_int_to_string takes an h3 integer representation and returns the string format.
}
