% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h3_string_to_int}
\alias{h3_string_to_int}
\title{Convert an H3 index string to 64 bit integer}
\usage{
h3_string_to_int(index)
}
\arguments{
\item{index}{h3 index as a hex representation character vector. See \code{\link{getIndexFromCoords}}}
}
\description{
h3_string_to_int takes an h3 index represented as a hexadecimal string and returns the integer format.
}
