% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_placekey.R
\name{is_placekey}
\alias{is_placekey}
\title{Check if a placekey is valid}
\usage{
is_placekey(placekey)
}
\arguments{
\item{placekey}{A Placekey.}
}
\description{
Check if a placekey is valid
}
\examples{
is_placekey("@b9f-q4k-d7r")
is_placekey("123@b9f-q4k-d7r")
}
