% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_placekey.R
\name{search_placekey}
\alias{search_placekey}
\title{Fetch place attributes by placekey}
\usage{
search_placekey(placekey, key = Sys.getenv("SAFEGRAPH_SECRET"))
}
\arguments{
\item{placekey}{A vector of placekeys. Must not exceed 100 values.}

\item{key}{Your SafeGraph API key. Recommend to be stored as an environment variable called `SAFEKEY_SECRET`.}
}
\description{
Use the SafeGraph [Places API](https://docs.safegraph.com/reference#places-api) to fetch POI attributes for a given placekey.
}
\details{
The SafeGraph places API will only return up to 4000 rows. This utilize your $200 credit. Each subsequent placekey will cost $0.05. You will not be charged if you have not set up payment.

Currently pagination and rate limiting for this API endpoint are not implemented.

If you do not have an API key, generate one on SafeGraph's [website](https://shop.safegraph.com/api/).
}
