% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_placekey.R
\name{get_placekey}
\alias{get_placekey}
\title{Query the Placekey API}
\usage{
get_placekey(
  location_name = NULL,
  street_address = NULL,
  city = NULL,
  region = NULL,
  postal_code = NULL,
  iso_country_code = NULL,
  strict_address_match = FALSE,
  strict_name_match = FALSE,
  key = Sys.getenv("PLACEKEY_SECRET")
)
}
\description{
Queries the Placekey API and, if successful, returns a placekey.
}
