% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_placekeys.R
\name{get_placekeys}
\alias{get_placekeys}
\title{Get Placekeys using the bulk API}
\usage{
get_placekeys(
  location_name = NULL,
  street_address = NULL,
  city = NULL,
  region = NULL,
  postal_code = NULL,
  iso_country_code = NULL,
  strict_address_match = FALSE,
  strict_name_match = FALSE,
  key = Sys.getenv("PLACEKEY_SECRET"),
  ...
)
}
\arguments{
\item{location_name}{The name of the place.}

\item{street_address}{The streed address of the place.}

\item{city}{The city where the place is located.}

\item{region}{The second-level administrative region below nation for the place. In the US, this is the state.}

\item{postal_code}{The postal code for the place.}

\item{iso_country_code}{The ISO 2-letter Country Code for the place. Currently may only be US.}

\item{strict_address_match}{If set to `TRUE`, a Placekey is only returned if all fields identify the place as being at the exact address specified. If set to `FALSE`, the Placekey returned may ignore unit/apartment/floor information. Optional. Default `FALSE`.}

\item{strict_name_match}{If set to `TRUE`, a Placekey is only returned if all fields identify the POI as having the exact name specified. Optional. Default `FALSE`.}

\item{...}{Unused; allows \code{purrr::pmap_chr} to call \code{get_placekey} with columns not used in the Placekey call.}

\item{longitude}{Longitude in decimal degrees.}

\item{latitude}{Latitude in decimal degrees.}
}
\description{
Get Placekeys using the bulk API
}
