% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{h3_int_to_string}
\alias{h3_int_to_string}
\title{Returns string format of an H3 integer representation}
\usage{
h3_int_to_string(index)
}
\arguments{
\item{index}{as an integer representation.}
}
\description{
Returns string format of an H3 integer representation
}
