% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placekey.R
\docType{placekey}
\name{placekey}
\alias{placekey}
\title{placekey: A package for computating the notorious bar statistic.}
\description{
An R library to interact with plackeys. The package provides access to the placekey.io API, conversion between Uber's H3 index, and other general placekey actions.
}
